Fix Supervisor.Spec deprecation
This really makes me wish for ember-cli-update-style diffs.
