+++ lib/adventure_registrations.ex
@@ -1,32 +1,9 @@
 defmodule AdventureRegistrations do
-  use Application
+  @moduledoc """
+  Phoenixnew keeps the contexts that define your domain
+  and business logic.
 
-  # See http://elixir-lang.org/docs/stable/elixir/Application.html
-  # for more information on OTP Applications
-  def start(_type, _args) do
-    import Supervisor.Spec, warn: false
-
-    children = [
-      # Start the PubSub system
-      {Phoenix.PubSub, name: AdventureRegistrations.PubSub},
-      # Start the endpoint when the application starts
-      supervisor(AdventureRegistrationsWeb.Endpoint, []),
-      # Start the Ecto repository
-      supervisor(AdventureRegistrations.Repo, [])
-      # Here you could define other workers and supervisors as children
-      # worker(AdventureRegistrations.Worker, [arg1, arg2, arg3]),
-    ]
-
-    # See http://elixir-lang.org/docs/stable/elixir/Supervisor.html
-    # for other strategies and supported options
-    opts = [strategy: :one_for_one, name: AdventureRegistrations.Supervisor]
-    Supervisor.start_link(children, opts)
-  end
-
-  # Tell Phoenix to update the endpoint configuration
-  # whenever the application is updated.
-  def config_change(changed, _new, removed) do
-    AdventureRegistrationsWeb.Endpoint.config_change(changed, removed)
-    :ok
-  end
+  Contexts are also responsible for managing your data, regardless
+  if it comes from the database, an external API or others.
+  """
 end
