+++ lib/adventure_registrations/application.ex
@@ -1,0 +1,34 @@
+defmodule AdventureRegistrations.Application do
+  # See https://hexdocs.pm/elixir/Application.html
+  # for more information on OTP Applications
+  @moduledoc false
+
+  use Application
+
+  @impl true
+  def start(_type, _args) do
+    children = [
+      # Start the Ecto repository
+      AdventureRegistrations.Repo,
+      # Start the PubSub system
+      {Phoenix.PubSub, name: AdventureRegistrations.PubSub},
+      # Start the Endpoint (http/https)
+      AdventureRegistrationsWeb.Endpoint
+      # Start a worker by calling: AdventureRegistrations.Worker.start_link(arg)
+      # {AdventureRegistrations.Worker, arg}
+    ]
+
+    # See https://hexdocs.pm/elixir/Supervisor.html
+    # for other strategies and supported options
+    opts = [strategy: :one_for_one, name: AdventureRegistrations.Supervisor]
+    Supervisor.start_link(children, opts)
+  end
+
+  # Tell Phoenix to update the endpoint configuration
+  # whenever the application is updated.
+  @impl true
+  def config_change(changed, _new, removed) do
+    AdventureRegistrationsWeb.Endpoint.config_change(changed, removed)
+    :ok
+  end
+end
